moved button click test to dashboard (test passing now)
